Allow `op.call` to work with async ops
Hi! Not sure how much we broke with this one change, but - we're trying to leverage this new API, which is very promising, and we had a "this coroutine is never awaited" error, which we traced to this place in your code. Hopefully this gives you an idea of how to provide a proper fix, and it helps you troubleshoot this problem faster.